update Thursday with Y

+++ src/main/java/com/jb/Project_coupon_3/controllers/CustomerController.java
@@ -5,6 +5,9 @@ import com.jb.Project_coupon_3.models.Category;
 import com.jb.Project_coupon_3.models.Coupon;
 import com.jb.Project_coupon_3.models.Customer;
 import com.jb.Project_coupon_3.services.CustomerService;
+import com.jb.Project_coupon_3.services.TokenService;
+import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,10 +16,14 @@ import java.util.List;
 import java.util.Set;
 
 @RestController
-@RequestMapping("customer")
+@RequestMapping("/customer")
 @CrossOrigin("*")
 public class CustomerController extends ClientController{
     CustomerService customerService;
+    TokenService tokenService;
+
+    @Autowired
+    HttpServletRequest request;
 
     public CustomerController(CustomerService customerService) {
         this.customerService = customerService;
@@ -29,37 +36,47 @@ public class CustomerController extends ClientController{
 
     @PostMapping("purchase")
     @ResponseStatus(HttpStatus.ACCEPTED)
-    public void couponPurchase(@RequestParam int couponId, @RequestParam int customerId) throws CouponSystemException {
+    public void couponPurchase(@RequestParam int couponId) throws CouponSystemException {
+        int customerId = tokenService.getId(request, "customer");
         customerService.couponPurchase(couponId, customerId);
     }
     @GetMapping("{customerId}")
     @ResponseStatus(HttpStatus.FOUND)
-    public Customer getOneCustomer(@PathVariable int customerId){
+    public Customer getOneCustomer(@PathVariable int customerId) throws CouponSystemException {
+        int customerTokenId = tokenService.getId(request, "customer");
+        if(customerTokenId != customerId)
+            throw new CouponSystemException("Please login with correct customer ID");
         return customerService.getOneCustomer(customerId);
     }
 
     @GetMapping("coupons")
     @ResponseStatus(HttpStatus.OK)
-    public Set<Coupon> getAllCustomerCoupons(@RequestParam int customerId){
+    public Set<Coupon> getAllCustomerCoupons(@RequestParam int customerId) throws CouponSystemException {
+        int customerTokenId = tokenService.getId(request, "customer");
+        if(customerTokenId != customerId)
+            throw new CouponSystemException("Please login with correct customer ID");
         return customerService.getAllCustomerCoupons(customerId);
     }
 
-    @GetMapping("coupons/category/{category}")
+    @GetMapping("coupons/category")
     @ResponseStatus(HttpStatus.OK)
-    public List<Coupon> getCouponsByCategory(@PathVariable Category category, @RequestParam int customerId){
+    public List<Coupon> getCouponsByCategory(@RequestParam Category category) throws CouponSystemException {
+        int customerId = tokenService.getId(request, "customer");
         return customerService.getCustomerCouponsByCategory(category, customerId);
     }
 
-    @GetMapping("coupons/price/{price}")
-    public List<Coupon> getCouponsByMaxPrice(@PathVariable double price,@RequestParam int customerId){
+    @GetMapping("coupons/price")
+    public List<Coupon> getCouponsByMaxPrice(@RequestParam double price) throws CouponSystemException {
+        int customerId = tokenService.getId(request, "customer");
         return customerService.getCouponsByMaxPrice(price, customerId);
     }
 
     //TODO - make sure only the customer that logged in can delete
     @DeleteMapping("{customerId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCustomer(@PathVariable int customerId){
-        customerService.deleteCustomer(customerId);
+    public void deleteCustomer(@PathVariable int customerId) throws CouponSystemException {
+        if(tokenService.getId(request, "customer") == customerId)
+            customerService.deleteCustomer(customerId);
     }
